COM-590 Fix Customer model and customer export tests
