+++ helpers/export.js
@@ -14,6 +14,7 @@ const UtilsHelper = require('./utils');
 const Bill = require('../models/Bill');
 const CreditNote = require('../models/CreditNote');
 const Contract = require('../models/Contract');
+const Customer = require('../models/Customer');
 const EventRepository = require('../repositories/EventRepository');
 
 const workingEventExportHeader = [
@@ -275,6 +276,79 @@ exports.exportContractHistory = async (startDate, endDate) => {
         ]);
       }
     }
+  }
+
+  return rows;
+};
+
+const getServicesNameList = (subscriptions) => {
+  let list = `${UtilsHelper.getLastVersion(subscriptions[0].service.versions, 'startDate').name}`;
+  if (subscriptions.length > 1) {
+    for (const sub of subscriptions.slice(1)) {
+      list = list.concat(`\r\n ${UtilsHelper.getLastVersion(sub.service.versions, 'startDate').name}`);
+    }
+  }
+  return list;
+};
+
+const customerExportHeader = [
+  'Email',
+  'Titre',
+  'Nom',
+  'Prenom',
+  'Date de naissance',
+  'Adresse',
+  'Environnement',
+  'Objectifs',
+  'Autres',
+  'Référente',
+  'Nom associé au compte bancaire',
+  'IBAN',
+  'BIC',
+  'RUM',
+  'Date de signature du mandat',
+  'Nombre de souscriptions',
+  'Souscriptions',
+  'Nombre de financements',
+  'Date de création',
+];
+
+exports.exportCustomers = async () => {
+  const customers = await Customer.find().populate('subscriptions.service').lean();
+  console.error(customers[0].subscriptions);
+  const rows = [customerExportHeader];
+
+  for (const cus of customers) {
+    const birthDate = get(cus, 'identity.birthDate');
+    const lastname = get(cus, 'identity.lastname');
+    const mandates = get(cus, 'payment.mandates');
+    const lastMandate = mandates ? UtilsHelper.getLastVersion(mandates, 'createdAt') || {} : {};
+    const signedAt = lastMandate.signedAt ? moment(lastMandate.signedAt).format('DD/MM/YYYY') : '';
+    const subscriptionsCount = get(cus, 'subscriptions.length') || 0;
+
+    const cells = [
+      cus.email || '',
+      get(cus, 'identity.title') || '',
+      lastname ? lastname.toUpperCase() : '',
+      get(cus, 'identity.firstname') || '',
+      birthDate ? moment(birthDate).format('DD/MM/YYYY') : '',
+      get(cus, 'contact.address.fullAddress') || '',
+      get(cus, 'followUp.environment') || '',
+      get(cus, 'followUp.objectives') || '',
+      get(cus, 'followUp.misc') || '',
+      get(cus, 'followUp.referent') || '',
+      get(cus, 'payment.bankAccountOwner') || '',
+      get(cus, 'payment.iban') || '',
+      get(cus, 'payment.bic') || '',
+      lastMandate.rum || '',
+      signedAt,
+      subscriptionsCount,
+      subscriptionsCount ? getServicesNameList(cus.subscriptions) : '',
+      get(cus, 'fundings.length') || 0,
+      cus.createdAt ? moment(cus.createdAt).format('DD/MM/YYYY') : '',
+    ];
+
+    rows.push(cells);
   }
 
   return rows;
